backend/requirements.txt
Django
djangorestframework==3.15.2
django-cors-headers==4.3.1
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.126.0
requests==2.31.0
python-dotenv==1.0.1
whitenoise==6.6.0
gunicorn==22.0.0
#psycopg2-binary==2.9.9
dj-database-url==2.1.0
